refactor: restructure folder structure
Folder structure was changed in order to differentiate better beween the namespaces

BREAKING CHANGE: File include paths and method names changed.
