+++ lib/caldav.dart
@@ -1,5 +1,4 @@
 library caldav;
 
-export 'src/client.dart';
-export 'src/objects.dart';
-export 'src/types.dart';
+export './src/caldav/calendar.dart';
+export './src/caldav/client.dart';
