+++ lib/src/caldav/calendar.dart
@@ -1,0 +1,11 @@
+class CalDavCalendar {
+  String path;
+  String displayName;
+
+  CalDavCalendar(this.path, this.displayName);
+
+  @override
+  String toString() {
+    return '$displayName (path: $path)';
+  }
+}
