--- lib/src/types.dart
@@ -1,5 +1,0 @@
-import 'objects.dart';
-
-bool isHrefProp(o) {
-  return o == new WebDavProp('href');
-}
