+++ lib/src/webdav/core/dav_resource.dart
@@ -1,0 +1,13 @@
+import '../properties/resourcetype.dart';
+
+class DavResource {
+  WebDavResourceType resourceType;
+  String path;
+
+  DavResource(this.path, this.resourceType);
+
+  @override
+  String toString() {
+    return 'WebDavResource{path: $path, resourceType: $resourceType}';
+  }
+}
