+++ lib/src/webdav/core/webdav_parser.dart
@@ -1,0 +1,8 @@
+import '../../core/parser.dart';
+
+abstract class WebDavParser<T> extends Parser<T> {
+  @override
+  String getNodeNamespace() {
+    return 'DAV:';
+  }
+}
