+++ lib/src/webdav/element/status.dart
@@ -1,0 +1,12 @@
+/// <status> element described in RFC 4918
+class WebDavStatus {
+  String status;
+
+  /// Holds a single HTTP status-line.
+  WebDavStatus(this.status);
+
+  @override
+  String toString() {
+    return this.status;
+  }
+}
