+++ lib/src/webdav/exception/webdav_exception.dart
@@ -1,0 +1,5 @@
+class WebDavException implements Exception {
+  String cause;
+
+  WebDavException(this.cause);
+}
