+++ test/src/xml/parsers_test.dart
@@ -1,10 +1,13 @@
-import 'package:caldav/src/xml/webdavresponse_parser.dart';
+import 'package:caldav/src/webdav/parser/webdavresponse_parser.dart';
 import 'package:test/test.dart';
 import 'package:xml/xml.dart' as xml;
 
 import '../../_fixtures/caldav_data.dart' as data;
-import 'package:caldav/src/objects.dart';
-import 'package:caldav/src/types.dart' as types;
+import 'package:caldav/src/webdav/element/_elements.dart';
+
+bool isHrefProp(o) {
+  return o == new WebDavProp('href');
+}
 
 void main() {
   test('parse XML into objects', () {
@@ -20,11 +23,12 @@ void main() {
     expect(result.elementAt(0).propStats.length, 1);
 
     var propStat = result.elementAt(0).propStats.elementAt(0);
+    expect(propStat.status.status, 'HTTP/1.1 200 OK');
     var currentUserPrincipal = propStat.props.firstWhere((prop) {
       return prop.name == 'current-user-principal' && prop.namespace == 'DAV:';
     });
     var href = currentUserPrincipal.value as List<WebDavProp>;
-    var hrefObj = href.firstWhere(types.isHrefProp);
+    var hrefObj = href.firstWhere(isHrefProp);
     expect(hrefObj.value, '/remote.php/caldav/principals/saitho/');
   });
